Revert "[savedObjects] change the number mappings to be 'long'"

+++ src/plugins/kibana/public/visualize/saved_visualizations/_saved_vis.js
@@ -48,7 +48,7 @@ define(function (require) {
       uiStateJSON: 'string',
       description: 'string',
       savedSearchId: 'string',
-      version: 'long'
+      version: 'integer'
     };
 
     SavedVis.searchSource = true;
